ref: extract PhelConfig deprecated shims into a trait
Move the 26 deprecated set*()/use*Layout() backward-compat shims out of
PhelConfig into a PhelConfigDeprecations trait (799 -> 624 lines). Public
surface and behavior are unchanged: the shims still resolve on PhelConfig
via the trait and keep their #[Deprecated] markers, so the canonical
with*() API is no longer buried under ~175 lines of aliases.
