+++ src/php/Config/CLAUDE.md
@@ -6,9 +6,9 @@ Pure data/model layer defining configuration structure for Phel projects. No Gac
 
 ### PhelConfig
 
-Immutable `final readonly class` (~650 LOC). Every mutation returns a new instance via `with*()` methods (directory, layout, build, export, cache, flag setters); `withOptimizationLevel(int)` (key `optimization-level`, clamped >= 0) sets the compiler optimization level consumed by Build/Run (REPL/nREPL ignore it); `forProject(ProjectLayout = Flat, string $mainNamespace = '')` convenience constructor; `validate(): list<string>` (empty if valid); `jsonSerialize()`.
+Immutable `final readonly class` (~620 LOC). Every mutation returns a new instance via `with*()` methods (directory, layout, build, export, cache, flag setters); `withOptimizationLevel(int)` (key `optimization-level`, clamped >= 0) sets the compiler optimization level consumed by Build/Run (REPL/nREPL ignore it); `forProject(ProjectLayout = Flat, string $mainNamespace = '')` convenience constructor; `validate(): list<string>` (empty if valid); `jsonSerialize()`.
 
-**Deprecated (since 0.37)**: `setX()` and `useLayout()`/`useNestedLayout()`/`useFlatLayout()` shim to `with*()` counterparts (marked `#[Deprecated]`). Permanent backward-compatibility aliases, not scheduled for removal in a minor release; removal would require a major version bump.
+**Deprecated (since 0.37)**: `setX()` and `useLayout()`/`useNestedLayout()`/`useFlatLayout()` shim to `with*()` counterparts (marked `#[Deprecated]`), kept in the `PhelConfigDeprecations` trait so they do not clutter the canonical API. Permanent backward-compatibility aliases, not scheduled for removal in a minor release; removal would require a major version bump.
 
 ### PhelBuildConfig / PhelExportConfig
